refactor: extract components for SRP (#18)
Split chat-header, chat-message, and main-page into smaller,
single-responsibility components: mode-toggle, user-message,
assistant-message, and message-list. 22 new tests added.
