+++ client/components/chat-header.js
@@ -1,5 +1,6 @@
 /** @typedef {import("../../shared/types.js").Mode} Mode */
 
+import "./mode-toggle.js";
 import { LitElement, css } from "lit-element";
 import { html } from "lit-html";
 import { customElement } from "lit/decorators.js";
@@ -38,39 +39,6 @@ class ChatHeader extends LitElement {
                 font-size: 0.875rem;
                 color: var(--muted-foreground);
             }
-            .mode-toggle {
-                display: flex;
-                align-items: center;
-                gap: 0.75rem;
-                background: var(--secondary);
-                border-radius: 0.5rem;
-                padding: 0.25rem;
-                border: 1px solid var(--border);
-            }
-            .mode-btn {
-                padding: 0.375rem 0.75rem;
-                font-size: 0.75rem;
-                font-weight: 500;
-                border-radius: 0.375rem;
-                border: none;
-                cursor: pointer;
-                transition:
-                    all var(--transition-speed),
-                    background-color var(--accent-transition-speed);
-                background: transparent;
-                line-height: 1rem;
-            }
-            .mode-btn.active {
-                color: white;
-                box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
-                background: var(--accent);
-            }
-            .mode-btn.inactive {
-                color: var(--muted-foreground);
-            }
-            .mode-btn.inactive:hover {
-                color: var(--foreground);
-            }
         `,
     ];
 
@@ -91,31 +59,22 @@ class ChatHeader extends LitElement {
                     <span class="title">Pathfinder 2e</span>
                     <span class="subtitle">Rules Assistant</span>
                 </div>
-                <div class="mode-toggle">
-                    <button
-                        @click=${() => this.setMode("player")}
-                        class="mode-btn ${this.mode === "player" ? "active" : "inactive"}"
-                    >
-                        ⚔️ Player Mode
-                    </button>
-                    <button
-                        @click=${() => this.setMode("gm")}
-                        class="mode-btn ${this.mode === "gm" ? "active" : "inactive"}"
-                    >
-                        📜 GM Mode
-                    </button>
-                </div>
+                <mode-toggle .mode=${this.mode} @mode-change=${this.handleModeChange}></mode-toggle>
             </header>
         `;
     }
 
     /**
-     * @param {Mode} mode
+     * @param {CustomEvent<{ mode: Mode }>} e
      */
-    setMode(mode) {
-        this.mode = mode;
+    handleModeChange(e) {
+        this.mode = e.detail.mode;
         this.dispatchEvent(
-            new CustomEvent("mode-change", { detail: { mode }, bubbles: true, composed: true }),
+            new CustomEvent("mode-change", {
+                detail: { mode: e.detail.mode },
+                bubbles: true,
+                composed: true,
+            }),
         );
     }
 }
