+++ client/pages/main-page.js
@@ -1,10 +1,9 @@
 import "../components/chat-header.js";
 import "../components/chat-input.js";
-import "../components/chat-message.js";
 import "../components/chat-sidebar.js";
-import "https://esm.sh/@shoelace-style/shoelace@2.20.1/dist/components/spinner/spinner.js?deps=lit@3.3.2";
+import "../components/message-list.js";
 import { LitElement, css } from "lit-element";
-import { html, nothing } from "lit-html";
+import { html } from "lit-html";
 import { customElement } from "lit/decorators.js";
 
 import { baseStyles } from "../styles/base-styles.js";
@@ -48,33 +47,6 @@ class MainPage extends LitElement {
                 display: flex;
                 flex-direction: column;
             }
-            .messages {
-                flex: 1;
-                overflow-y: auto;
-                padding: 1.5rem;
-            }
-            .messages > * + * {
-                margin-top: 1.5rem;
-            }
-            .messages::-webkit-scrollbar {
-                width: 6px;
-            }
-            .messages::-webkit-scrollbar-track {
-                background: transparent;
-            }
-            .messages::-webkit-scrollbar-thumb {
-                background: var(--border-lighter);
-                border-radius: 3px;
-            }
-            .loading {
-                display: flex;
-                align-items: center;
-                gap: 0.5rem;
-                color: var(--muted-foreground);
-            }
-            .loading-text {
-                font-size: 0.875rem;
-            }
         `,
     ];
 
@@ -102,33 +74,6 @@ class MainPage extends LitElement {
         this.sidebarExpanded = true;
     }
 
-    scrollToBottom() {
-        if (!this.shadowRoot) {
-            return;
-        }
-        const container = this.shadowRoot.querySelector(".messages");
-        if (!container) {
-            return;
-        }
-        container.scrollTo({
-            top: container.scrollHeight,
-            behavior: "smooth",
-        });
-    }
-
-    firstUpdated() {
-        this.scrollToBottom();
-    }
-
-    /**
-     * @param {Map<string, unknown>} changedProperties
-     */
-    updated(changedProperties) {
-        if (changedProperties.has("messages")) {
-            this.scrollToBottom();
-        }
-    }
-
     render() {
         return html`
             <div class="app" data-mode=${this.mode}>
@@ -146,19 +91,10 @@ class MainPage extends LitElement {
                         .mode=${this.mode}
                         @mode-change=${this.handleModeChange}
                     ></chat-header>
-                    <div class="messages">
-                        ${this.messages.map(
-                            (msg) => html` <chat-message .message=${msg}></chat-message> `,
-                        )}
-                        ${this.loading
-                            ? html`
-                                  <div class="loading">
-                                      <sl-spinner style="font-size: 1rem;"></sl-spinner>
-                                      <span class="loading-text">Thinking...</span>
-                                  </div>
-                              `
-                            : nothing}
-                    </div>
+                    <message-list
+                        .messages=${this.messages}
+                        .loading=${this.loading}
+                    ></message-list>
                     <chat-input
                         .mode=${this.mode}
                         @send-message=${this.handleSendMessage}
